Fix #945

+++ zp-core/zp-extensions/galleryArticles.php
@@ -42,8 +42,10 @@ class galleryArticles {
 			setOptionDefault('galleryArticles_image_text', getAllTranslations($text));
 			setOptionDefault('galleryArticles_size', 80);
 			setOptionDefault('galleryArticles_protected', 0);
-			cacheManager::deleteThemeCacheSizes('galleryArticles');
-			cacheManager::addThemeCacheSize('galleryArticles', getOption('galleryArticles_size'), NULL, NULL, NULL, NULL, NULL, NULL, false, NULL, NULL, NULL);
+			if (class_exists('cacheManager')) {
+				cacheManager::deleteThemeCacheSizes('galleryArticles');
+				cacheManager::addThemeCacheSize('galleryArticles', getOption('galleryArticles_size'), NULL, NULL, NULL, NULL, NULL, NULL, false, NULL, NULL, NULL);
+			}
 		}
 	}
